Enable binary cache usage
This adds a new Make task `cache` that will build and push a machine
configuration into the cache (requires an auth token).

This configures pulling to use my binary cache, which is public and
doesn't require any auth tokens. I only push my NixOS configs there
which are already all public and contain no secrets.

Cachix has to be configured in bootstrap0 because it requires a full
complete Nix run before its active (not active until the next run). So
by doing it in bootstrap0, our bootstrap uses the cache which is the
custom part.